+++ entrypoints/background.ts
@@ -1,6 +1,3 @@
-import { Runtime, Tabs } from "wxt/browser"
-import axios from "axios"
-
 import {
   type ExtractingCheckMessage,
   type InjectingCheckMessage,
@@ -11,7 +8,7 @@ import {
 } from "@/utils/messages"
 import { Event } from "@/utils/event"
 import { isSupportedTab } from "@/utils/browser_tabs"
-import { Workflow } from "@/utils/workflow"
+import { topLevelBackgroundMsgListener } from "@/utils/trigger_workflow"
 
 // Tabs may not have our content script injected if they were already open
 // before the plugin itself was installed. This array tracks the different
@@ -158,77 +155,10 @@ const syncRecordingState = async (tabId: number) => {
   }
 }
 
-const fetchWorkflow = async (tabURL: string): Promise<Workflow> => {
-  // TODO: Support enqueue of multiple workflows
-  const workflowIdsToEnqueue = new URL(tabURL).searchParams.getAll(
-    "fluidicWorkflowIds[]"
-  )
-  const [workflowId] = workflowIdsToEnqueue.map((s) => String(s).trim())
-  if (!workflowId) {
-    throw new Error(
-      "No workflowId found in tab url query params, url: " + tabURL
-    )
-  }
-
-  // Query workflowId as Workflow
-  // TODO: Replace with env-based api target OR propagate target in evt payload
-  const response = await axios.get(
-    `http://localhost:80/api/workflows/${workflowId}`
-  )
-  const workflow: Workflow = response.data
-  if (!workflow) {
-    throw new Error(`No workflow found for id ${workflowId} from response`)
-  }
-
-  return workflow
-}
-
 const syncTab = async (tabId: number) => {
   await syncExtractingState(tabId)
   await syncInjectingState(tabId)
   await syncRecordingState(tabId)
-}
-
-const topLevelBackgroundMsgListener = <M extends Message>(
-  message: M,
-  sender?: Runtime.MessageSender
-) => {
-  ;(async () => {
-    switch (message.event) {
-      case Event.TRIGGER_WORKFLOW_CHECK: {
-        const tab = sender?.tab
-        const { id: tabId, windowId, url: tabURL } = sender?.tab ?? {}
-        if (!tab || !tabId || !tabURL)
-          throw new Error( // TRIGGER_WORKFLOW_QUERY should not be sent until landing page has loaded, button has been injected, and tapped to open extension
-            `No tab id ${tabId} or tab url ${JSON.stringify(sender?.tab)} found from evt sender`
-          )
-
-        // Open side panel programatically — nesting or wrapping in any handler (without injecting a button from chrome runtime trigger by client) will lead to Error — user gesture required
-        // @ts-ignore
-        browser.sidePanel.open({ windowId })
-
-        // Query workflow data and begin internal execution of steps sequentially
-        const workflow = await fetchWorkflow(tabURL)
-        const tabs = await browser.tabs.query({
-          active: true,
-          currentWindow: true,
-        })
-        const anchorHeadlessTab = tabs.find((t) =>
-          t.url?.includes("headless/home?fluidicWorkflowIds[]")
-        )
-        if (!anchorHeadlessTab || !anchorHeadlessTab.id)
-          throw new Error(
-            "No headless found from tabs query for URL headless/home?fluidicWorkflowIds[]"
-          )
-        await sendTab(anchorHeadlessTab.id, {
-          event: Event.TRIGGER_WORKFLOW_START,
-          payload: { workflow },
-        })
-        removeMessageListener(topLevelBackgroundMsgListener)
-        return
-      }
-    }
-  })()
 }
 
 const definition: ReturnType<typeof defineBackground> = defineBackground(() => {
@@ -257,8 +187,7 @@ const definition: ReturnType<typeof defineBackground> = defineBackground(() => {
     await syncTab(tabId)
   })
 
-  // chrome.runtime.onMessage listeners must be set up in background service via top-level onDefineBackground;
-  // onMessage/onMessageExternal is undefined in content scripts and within tab listeners above.
+  // top-level background listener to init flows within content script inject components; must be in background.ts and async iife wrapped to avoid 'Error - user gesture required'
   addMessageListener(topLevelBackgroundMsgListener)
 })
